Windows UDP support.
This implements the basic UDP functionality for Windows (required
for ZeroTier for example).  We have also introduced a UDP test suite
to validate that this actually works.  While here a few Windows
compilation warnings / nits were fixed.

+++ tests/trantest.h
@@ -183,7 +183,7 @@ trantest_send_recv_large(trantest *tt)
 		size = 1024 * 128; // bigger than any transport segment
 		So((data = nng_alloc(size)) != NULL);
 
-		for (int i = 0; i < size; i++) {
+		for (int i = 0; (size_t) i < size; i++) {
 			data[i] = nni_random() & 0xff;
 		}
